move row actions to saparate component
